04_gcp_components/components/spark_component/requirements.txt
pyspark
google-cloud-dataproc
google-cloud-storage
